challenge_backend_2025/requirements.txt
asgiref==3.8.1
attrs==25.3.0
cffi==1.17.1
colorama==0.4.6
cryptography==43.0.1
Django==5.2.1
django-dotenv==1.4.2
djangorestframework==3.16.0
drf-spectacular==0.28.0
exceptiongroup==1.3.0
gunicorn==23.0.0
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
packaging==25.0
pillow==11.2.1
pluggy==1.6.0
pycparser==2.22
pytest==8.3.5
pytest-django==4.11.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.25.1
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.13.2
tzdata==2023.2
uritemplate==4.1.1
